requirements.txt
fastapi
uvicorn
authlib
google-cloud-storage
google-cloud-firestore
python-dotenv
jinja2